Use SDK info to find operation from method name (#70)
*Issue #, if available:*
https://github.com/awslabs/iam-policy-autopilot/issues/66

*Description of changes:*

We now look up method names in the service reference when we detect the
SDK type is boto3 (i.e., when the input files are Python). This fixes an
issue where we wrongly converted method names to operation names via
`convert_case`'s `to_case(Case::PascalCase)` before, which fails for
operations like `ModifyDBCluster`, where the snake_case method name is
`modify_db_cluster` and the conversion results in `ModifyDbCluster`.
For other SDKs we use the extracted method names as-is as the operation
name. The conversion to PascalCase, also changed `ModifyDBCluster` to
`ModifyDbCluster`, leading to the same error for Go, Typescript and
Javascript.

I introduced `SdkType::Boto3` to follow the data the service reference
provides, such as
```
{
      "Name" : "rds",
      "Method" : "modify_db_cluster",
      "Package" : "Boto3"
}
```


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.